feat(j-s): fetch parent case files (#7739)
* feat(j-s): add component to show files of parent case

* refator(j-s): make onRemoveClick optional

* feat(j-s): show files of parent case

- Moved StepFiveForm.tsx code into StepFive.tsx

* test(j-s): add tests for displaying files of parent case

* refactor(j-s): move CaseFilesForm.tsx to CaseFiles.tsx

* Update apps/judicial-system/web-e2e/src/integration/Prosecutor/InvestigationCases/caseFiles.spec.ts


* Update apps/judicial-system/web-e2e/src/integration/Prosecutor/RestrictionCases/caseFiles.spec.ts


* fix(j-s): fix casing for contentful id

* fix(j-s): fetch parent case files in case query
