+++ apps/judicial-system/web-e2e/src/integration/Prosecutor/RestrictionCases/caseFiles.spec.ts
@@ -6,6 +6,7 @@ import {
   makeCourt,
   makeProsecutor,
   intercept,
+  makeCaseFile,
 } from '../../../utils'
 
 describe(`${STEP_FIVE_ROUTE}/:id`, () => {
@@ -25,5 +26,21 @@ describe(`${STEP_FIVE_ROUTE}/:id`, () => {
     intercept(caseDataAddition)
 
     cy.get('[name=fileUpload]').attachFile('lorem-ipsum.txt')
+  })
+
+  it('should list of files for parent case', () => {
+    const file1 = makeCaseFile('file1', 'file1')
+    const file2 = makeCaseFile('file2', 'file2')
+    const caseData = makeRestrictionCase()
+    const caseDataAddition: Case = {
+      ...caseData,
+      court: makeCourt(),
+      parentCase: { ...makeRestrictionCase(), caseFiles: [file1, file2] },
+    }
+
+    intercept(caseDataAddition)
+
+    cy.get('button[aria-controls="parentCaseFiles"]').click()
+    cy.get('#parentCaseFiles').children().should('have.length', 2)
   })
 })
