+++ apps/judicial-system/web/src/components/index.ts
@@ -30,3 +30,4 @@ export { default as SkipToContent } from './SkipToContent/SkipToContent'
 export { default as HideableText } from './HideableText/HideableText'
 export { default as ConclusionDraft } from './ConclusionDraft/ConclusionDraft'
 export { default as InfoBox } from './InfoBox/InfoBox'
+export { default as ParentCaseFiles } from './ParentCaseFiles/ParentCaseFiles'
